+++ packages/terra-application-navigation/src/common/_PopupMenu.jsx
@@ -16,6 +16,7 @@ import {
 
 import PopupMenuListItem from './_PopupMenuListItem';
 import { userConfigPropType } from '../utils/propTypes';
+import { logoutUtilityItemId } from '../utils/helpers';
 import PopupMenuUser from './_PopupMenuUser';
 
 import styles from './PopupMenu.module.scss';
@@ -31,6 +32,10 @@ const propTypes = {
    * The text to be assigned to the button in the footer.
    */
   footerText: PropTypes.string,
+  /**
+   * An id used to unique identify items
+   */
+  id: PropTypes.string,
   /**
    * Callback for when the footer item is selected.
    */
@@ -59,6 +64,10 @@ const propTypes = {
      * Key used as react key, and returned in the onSelect
      */
     key: PropTypes.string.isRequired,
+    /**
+     * The id for the menuItem.
+     */
+    id: PropTypes.string,
     /**
      * Object to be returned in the onSelect
      */
@@ -102,7 +111,7 @@ const defaultProps = {
 };
 
 const PopupMenu = ({
-  title, footerText, onSelectFooterItem, onSelectMenuItem, customContent, userConfig, menuItems, isHeightBounded, showSelections, role,
+  title, footerText, id, onSelectFooterItem, onSelectMenuItem, customContent, userConfig, menuItems, isHeightBounded, showSelections, role,
 }) => {
   const listRef = useRef();
   const buttonRef = useRef();
@@ -165,6 +174,7 @@ const PopupMenu = ({
   if (onSelectFooterItem) {
     endContent = (
       <Button
+        id={id && logoutUtilityItemId(id)}
         text={footerText}
         onClick={onSelectFooterItem}
         onKeyDown={handleButtonKeyDown}
@@ -193,6 +203,7 @@ const PopupMenu = ({
         <ul className={cx('utility-list')} aria-label={title} ref={listRef} role={role} tabIndex="0" onKeyDown={handleKeyDown}>
           {menuItems.map(item => (
             <PopupMenuListItem
+              id={item.id || undefined}
               key={item.key}
               onSelect={onSelectMenuItem && onSelectMenuItem.bind(null, item.key, item.metaData)}
               icon={item.icon}