+++ packages/terra-application-navigation/src/header/_Header.jsx
@@ -24,6 +24,10 @@ const propTypes = {
    */
   navigationItems: navigationItemsPropType,
   /**
+   * The base id used to generate ids of navigation, utility, and extension items
+   */
+  id: PropTypes.string,
+  /**
    * A function to be executed for the render of each navigation item.
    */
   navigationRenderFunction: PropTypes.func,
@@ -107,6 +111,7 @@ const defaultProps = {
 const Header = ({
   titleConfig,
   navigationItems,
+  id,
   navigationRenderFunction,
   activeNavigationItemKey,
   onSelectNavigationItem,
@@ -151,6 +156,7 @@ const Header = ({
     return (
       <Tabs
         navigationItems={navigationItems}
+        id={id}
         activeTabKey={activeNavigationItemKey}
         onTabSelect={onSelectNavigationItem}
         notifications={notifications}
@@ -167,6 +173,7 @@ const Header = ({
     return (
       <Extensions
         extensionItems={extensionItems}
+        id={id}
         activeBreakpoint={activeBreakpoint}
         onSelect={onSelectExtensionItem}
         notifications={notifications}
@@ -183,6 +190,7 @@ const Header = ({
       <div className={cx('utilities-container')}>
         <UtilityMenuHeaderButton
           userConfig={userConfig}
+          id={id}
           onClick={onSelectUtilityButton}
           popupAnchorRef={utilityButtonPopupAnchorRef}
         />