+++ packages/terra-application-navigation/CHANGELOG.md
@@ -7,6 +7,9 @@
 * Changed
   * Updated to terra-functional-testing
   * Remove extraneous application bases from tests.
+  
+* Added
+  * Adding an id to ApplicationNavigation now causes navigation, extesion, and utility items to have a unique id.
 
 * Fixed
   * Prevent error when using focus trap with react-intl v5